requirements.txt
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
certifi==2025.8.3
click==8.2.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
markdown-it-py==4.0.0
mcp==1.13.1
mdurl==0.1.2
psutil==7.0.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
referencing==0.36.2
rich==14.1.0
rpds-py==0.27.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sse-starlette==3.0.2
starlette==0.47.3
typer==0.17.4
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.35.0
